Movendo o input para o app
