+++ src/App.js
@@ -1,6 +1,6 @@
-import styled from 'styled-components';
+import React from "react";
+import styled from "styled-components";
 import './App.css';
-import Inputs from './components/Inputs';
 
 const Main = styled.main`
   background-color: #e5ddd5;
@@ -11,12 +11,107 @@ const Main = styled.main`
   margin: 0 auto;
 `
 
-function App() {
-  return (
-    <Main>
-      <Inputs enviarMensagem={ (mensagem) => console.log(mensagem) } />
-    </Main>
-  );
+const Form = styled.form`
+  display: grid;
+  grid-template-columns: 2fr 7fr 1fr;
+  grid-gap: 1%;
+  padding: 1%;
+  max-width: 100%;
+`
+
+const Input = styled.input`
+  width: 100%;
+  padding: 8px;
+  border: 0;
+  border-radius: 8px;
+  font-size: 16px;
+`
+
+const Button = styled.button`
+  background-color: #efefef;
+  width: 100%;
+  padding: 8px;
+  border: 0;
+  border-radius: 8px;
+  font-size: 16px;
+  font-weight: bold;
+  cursor: pointer;
+  transition: color ease-in-out 300ms, background-color ease-in-out 300ms, transform ease-in-out 300ms;
+
+  &:hover {
+    background-color: #d9d8d8;
+    color: #000000cc
+  }
+
+  &:active {
+    transform: translateY(6px)
+  }
+`
+
+class App extends React.Component {
+
+  state = {
+    remetenteInput: "",
+    mensagemInput: "",
+    mensagems: []
+  }
+
+  onChangeRemetenteInput = (event) => {
+
+    this.setState({ remetenteInput: event.target.value })
+
+  }
+
+  onChangeMensagemInput = (event) => {
+
+    this.setState({ mensagemInput: event.target.value })
+
+  }
+
+  enviarMensagem = () => {
+
+    const mensagem = {
+      remetente: this.state.remetenteInput,
+      mensagem: this.state.mensagemInput
+    }
+
+    if( !mensagem.remetente || !mensagem.mensagem ) {
+      return alert("A mensagem precisa de um usuário e de um conteúdo")
+    }
+
+
+    const novasMensagens = this.state.mensagems
+    novasMensagens.push(mensagem)
+    this.setState({
+      remetenteInput: "",
+      mensagemInput: "",
+      mensagems: novasMensagens
+    })
+
+  }
+
+  render() {
+    return (
+      <Main>
+      <Form>
+        <Input 
+          value={ this.state.remetenteInput }
+          onChange={ this.onChangeRemetenteInput } 
+          placeholder="Usuário"
+        />
+        <Input 
+          value={ this.state.mensagemInput }
+          onChange={ this.onChangeMensagemInput }
+          placeholder="Mensagem"
+
+        />
+        <Button type="button" onClick={ this.enviarMensagem } >
+          Enviar
+        </Button>
+      </Form>
+      </Main>
+    )
+  }
 }
 
-export default App;
+export default App
